finished postFilm
