+++ controllers/controllerFilms.js
@@ -5,4 +5,16 @@ const getAllFilms = async (req, res) => {
 	res.status(200).json(films)
 }
 
-module.exports = { getAllFilms }
+const postFilm = async (req, res) => {
+	const films = new model(req.body)
+	try {
+		const guardado = await films.save()
+		res.status(200).json(guardado)
+	} catch (error) {
+		res.status(500).json({
+			error: error.mensaje,
+		})
+	}
+}
+
+module.exports = { getAllFilms, postFilm }
